Creates the first webserver dockerfile version
- Description:
Stablishes a initial dockerfile for the webserver service capable of building and running a 'hello world' dotnet webapi project.

- Details:
The dockerfile is divided in two steps. The firs one is based on the
dotnet sdk base image and is used to build and publish the application.
The artifacts resultant of this phase are stored on the build directory.

The second step, that uses a runtime only base image, exposes the
necessary network ports, copy the build artifacts from the previus step
results directory, define the url that should be watched by the web
server through a environment variable and starts the application.