Adds HTTPS support to the webserver dev container
- Description:
As the hello world application is ready to accept https connections from
the get-go, a self signed SSL certificate (generated through the dotnet
cli) was added to the dev container.

- Details:
Using the dotnet cli a self-signed SSL certificate was generated to
enable the dev container to respond through HTTPS. The cerfificate is
copied to the runtime container, get its file permissions adjusted for
kestrel to be able to read it, and some environment variables are set to
instruct kestrel where to look and wich password to use. Lastly, the new
container version also exposes the 5001 port.